chore: move Token and HighlightedToken
