+++ src/Phiki.php
@@ -6,6 +6,8 @@ use Phiki\Contracts\GrammarRepositoryInterface;
 use Phiki\Contracts\ThemeRepositoryInterface;
 use Phiki\Grammar\GrammarRepository;
 use Phiki\Grammar\ParsedGrammar;
+use Phiki\Theme\ThemeRepository;
+use Phiki\Theme\ThemeStyles;
 
 class Phiki
 {
